# Schedule fibers on external queue if all worker threads busy