@vasilmkd 2021-11-07
As an added optimization, you should be able to remove this line, because it makes the same check (which is not completely free in terms of performance as it does a read of a volatile variable), and we're making the bet that we're only scheduling on the same worker thread if every other thread is busy.

https://github.com/typelevel/cats-effect/blob/b2276e5a96766fa9df9a86999cc1be246a43bd06/core/jvm/src/main/scala/cats/effect/unsafe/WorkerThread.scala#L102