@otto-dev 2021-11-08
> you should be able to remove this line `pool.notifyParked(rnd)`

Thanks. Can I remove this even if I leave `rescheduleFiber` untouched, and only modify `rescheduleFiber`?