@vasilmkd 2021-11-08
Thanks for taking this on! The graphs looks very encouraging.

> Thanks. Can I remove this even if I leave rescheduleFiber untouched, and only modify rescheduleFiber?

To be completely safe, remove the line from `WorkerThread#schedule` and add it after the call to `schedule` in `WorkerThread#reschedule`. It would also be fun to try the benchmarks without adding the line back in `reschedule`.