@otto-dev 2021-11-08
First benchmark result is in (took 30 minutes for both runs, so I'm setting up the other benchmarks separately)