@otto-dev 2021-11-08
> we're making the bet that we're only scheduling on the same worker thread if every other thread is busy.

I would have said "we're making the bet that we're **_not_** scheduling on the same worker thread if **_all_** threads are busy. That's at least my intent. Because if all threads are busy, you can have a situation where none of them steal work. So scheduling on the same worker thread (rather than through the external queue) can lead to an unfair distribution of fibers in such a situation. So in that situation we schedule through the external queue.

So in that case, if I read the code correctly, we might have to leave that line in. Please do correct me if I'm missing something!

Running the second set of benchmarks now. Looking promising so far.