@vasilmkd 2021-11-08
I'm just going to comment, this change removes the thread affinity of the pool (at full loads, fibers will be jumping around the pool all the time). Which benchmarks exactly are showing peformance improvements?