@djspiewak 2021-11-08
I need to think about this more deeply, but at first glance it seems a little concerning. It seems like this would result in child fibers losing all affinity to worker threads, which is an optimization that is not well captured by the current benchmarks. Thread affinity is impactful because the OS kernel will *also* try to do its own version of the same optimization, which is to say, keeping threads affine to host CPUs as much as possible. These two optimizations together improve cache utilization considerably, since child fibers are very likely to access state from their parent(s), which in turn will likely be in cache on the parent's CPU but not on other CPUs.

On the other side of the spectrum, I'm not entirely convinced that this covers a real use-case. We have a nice artificial test which can reproduce this scenario, but even in that scenario, throughput is still optimal, and presumably any truly compute-bound situation is going to care more about throughput than fairness. In any fairness-relevant scenario, other actions (such as `async` completion and internal forking) would take place which trigger the pool to balance on its own, meaning that proactively `start`ing on the external queue is unnecessary contention, since stealing will ultimately happen on its own.

Those are just initial thoughts though. Need to ponder some more.