@otto-dev 2021-11-08
> Which benchmarks exactly are showing peformance improvements?

I think this is irrelevant, but here is a list of benchmarks that performed better by score:
- `AsyncBenchmark.race`
- `DeepBindBenchmark.async`
- `DeepBindBenchmark.delay`
- `ShallowBindBenchmark.async`
- `ShallowBindBenchmark.delay`
- `ShallowBindBenchmark.pure`
- `WorkStealingBenchmark.alloc`
- `WorkStealingBenchmark.scheduling`

All differences (better or worse) are within the margin of `error`, and the average score is 101.5% of the previous version. This PR is not meant to improve performance, and I'd argue that it doesn't.

> at full loads, fibers will be jumping around the pool all the time

If that is the case, it doesn't show in the benchmarks, which are supposed to create the highest load possible. This PR only has an effect on computation-bound applications that lack asynchronous components. 

If there is a particular benchmark you are concerned about, I'm happy to run it with more iterations.