@vasilmkd 2021-11-08
The `WorkStealingBenchmark.alloc` and `WorkStealingBenchmark.scheduling` call `cede` every other operation. This _only_ exercises `reschedule`, not `schedule`. The new code isn't being run on them, that's why the results are not changing. I doubt that they are improving because either, again, because no new code is being run.