@otto-dev 2021-11-08
Is there any benchmark that would highlight your concern, or should we add a new one?