@vasilmkd 2021-11-08
Do you have the scores also? Thanks.