@otto-dev 2021-11-08
It's in the `after` section: https://github.com/typelevel/cats-effect/pull/2514#issue-774926194