@otto-dev 2021-11-12
To conclude this PR, I ran the same benchmark, measuring throughput, not fairness, to confirm that the change in fairness does not come at the expense of throughput. To do so I removed the `createUnfairness` part. [Link to benchmark.](https://gist.github.com/otto-dev/c60383d4653436fe748164a7cf782fca). Also, `size` here is now referring to "Computation intensity" above.

Results:

**Before**
```
[info] Benchmark                       (size)   Mode  Cnt     Score      Error    Units
[info] ThroughputBenchmark.scheduling       1  thrpt   20  4503.731 ±  221.476  ops/min
[info] ThroughputBenchmark.scheduling      10  thrpt   20  5410.462 ± 1208.783  ops/min
[info] ThroughputBenchmark.scheduling     100  thrpt   20  5403.228 ±  160.312  ops/min
[info] ThroughputBenchmark.scheduling    1000  thrpt   20   881.516 ±    6.627  ops/min
[info] ThroughputBenchmark.scheduling   10000  thrpt   20    95.048 ±    0.372  ops/min
[info] ThroughputBenchmark.scheduling  100000  thrpt   20     9.345 ±    0.080  ops/min
```

**After**
```
[info] Benchmark                       (size)   Mode  Cnt     Score      Error    Units
[info] ThroughputBenchmark.scheduling       1  thrpt   20  4225.515 ±  185.137  ops/min
[info] ThroughputBenchmark.scheduling      10  thrpt   20  5385.358 ± 1125.201  ops/min
[info] ThroughputBenchmark.scheduling     100  thrpt   20  5776.103 ±  104.549  ops/min
[info] ThroughputBenchmark.scheduling    1000  thrpt   20   907.226 ±   15.629  ops/min
[info] ThroughputBenchmark.scheduling   10000  thrpt   20    98.211 ±    1.207  ops/min
[info] ThroughputBenchmark.scheduling  100000  thrpt   20     9.335 ±    0.158  ops/min
```

This was the last thing I was going to do for this PR. My conclusion:

The PR noticeably improves fairness for computation intensive applications while having no significant impact on throughput.

There is a slight fairness reduction for non-computation intensive applications. But that doesn't matter. If you have millions of "light" short-lived fibers, you are not going to see that. But if you have thousands of long-lived computation intensive fibers, you may (e.g. an AI driven trading bot).

That's it from me. I leave it to the maintainers to make a decision.